Abstract CommentForm from CommentEditor
